requirements.txt
tiktoken
pyautogen==0.2.16
scipy==1.12.0
tiktoken==0.6.0
langchain-community==0.0.27
pillow==10.2.0
openai==1.13.3